refactor(apil): move internal consolidated to dedicated file +
- move internal consolidated to dedicated file
- make pt ID optional on internal consolidated search ingestion endpoint

Ref eng-268
